aienginev2: src: fix dma iteration stepsize data type
shim dma and memtile dma modules have iteration stepsize greater than 16bits.
change data type to u32 in apis to handle this.


Acked-by: Nishad Saraf <nishad.saraf@xilinx.com>

Acked-by: Wendy Liang <wendy.liang@xilinx.com>

+++ XilinxProcessorIPLib/drivers/aienginev2/src/dma/xaie_dma.c
@@ -469,7 +469,7 @@ AieRC XAie_DmaSetMultiDimAddr(XAie_DmaDesc *DmaDesc, XAie_DmaTensor *Tensor,
 *		hardware.
 *
 ******************************************************************************/
-AieRC XAie_DmaSetBdIteration(XAie_DmaDesc *DmaDesc, u16 StepSize, u8 Wrap,
+AieRC XAie_DmaSetBdIteration(XAie_DmaDesc *DmaDesc, u32 StepSize, u8 Wrap,
 		u8 IterCurr)
 {
 	const XAie_DmaMod *DmaMod;
